change type of BasePacket.id to uint8_t

+++ Source/Eurobalise_Telegram/BasePacket.h
@@ -3,13 +3,14 @@
 #define BASEPACKET_H_INCLUDED
 
 #include <cstdlib>
+#include <cstdint>
 #include <memory>
 #include <vector>
 #include <iostream>
 
 struct BasePacket
 {
-    int id;
+    uint8_t id;
 
     BasePacket(int id1) : id(id1) {}
